requirements.txt
# Django Core
Django==5.1.3
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.15.2

# OAuth2
django-oauth-toolkit==3.0.1

# CORS
django-cors-headers==4.4.0

# Image Processing
Pillow==10.4.0

# Environment Variables
python-decouple==3.8

# Barcode Generation & Label Printing
python-barcode==0.15.1
reportlab==4.2.5

# API Documentation
drf-spectacular==0.27.2

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1

# Utilities
python-dateutil==2.9.0

# Celery for background tasks
celery==5.4.0
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Excel export
openpyxl==3.1.2

# PDF generation
weasyprint==62.1

# HTTP requests (for SMS)
requests==2.31.0

